+++ src/SoilMoistureSensorCalibratedPumpESP/Common.h
@@ -6,17 +6,17 @@
 
 extern const int ANALOG_MAX;
 
-extern long lastSerialOutputTime; // Milliseconds
+extern unsigned long lastSerialOutputTime; // Milliseconds
 extern long serialOutputIntervalInSeconds;
 
 extern bool isDebugMode;
 
-extern int loopNumber;
+extern unsigned long loopNumber;
 
 void serialPrintLoopHeader();
 void serialPrintLoopFooter();
 
-long secondsToMilliseconds(int seconds);
+unsigned long secondsToMilliseconds(int seconds);
 float millisecondsToSecondsWithDecimal(int milliseconds);
 
 long readLong(char* text, int startPosition, int digitCount);
